Troca de BigDecimal para double

+++ mongodb/catalogo-de-produtos/src/main/java/br/com/compasso/catalogodeprodutos/form/ProductForm.java
@@ -1,13 +1,8 @@
 package br.com.compasso.catalogodeprodutos.form;
 
-import java.math.BigDecimal;
-
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
-
-import org.springframework.data.mongodb.core.mapping.Field;
-import org.springframework.data.mongodb.core.mapping.FieldType;
 
 import br.com.compasso.catalogodeprodutos.model.Product;
 
@@ -40,11 +35,11 @@ public class ProductForm {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
